# Allow kore-rpc-booster to read additional options from KORE_RPC_OPTS