* like `kore-rpc` , `kore-rpc-booster` should read additional options from an environment variable.
* This PR chooses the same variable name as for `kore-rpc` for consistency and to avoid it getting too long.

Use case: testing optional optimisation in performance tests without having to modify the pytest framework or the `KoreServer` in `pyk`