@jberthold 2024-05-02
I ran a KEVM performance test using `--no-fallback-simplify` (that was the use case for this feature). Looks promising!

| Test                                                                                      | HOTFIX-allow-kore-rpc-booster-args-in-env-var time | master-602850806 time | (HOTFIX-allow-kore-rpc-booster-args-in-env-var/master-602850806) time 
|-------------------------------------------------------------------------------------------|----------------------------------------------------|-----------------------|-----------------------------------------------------------------------
| kontrol/test-allowchangestest-testfailallowchangestostorage-0-spec.k                      | 54.5                                               | 108.81                | 0.5008730815182428                                                    
| erc20/hkg/transferFrom-failure-2-spec.k                                                   | 49.75                                              | 74.44                 | 0.6683234819989253                                                    
| erc20/hkg/approve-spec.k                                                                  | 39.09                                              | 56.45                 | 0.6924712134632418                                                    
| erc20/ds/transfer-failure-1-a-spec.k                                                      | 50.91                                              | 72.11                 | 0.7060047150187213                                                    
| mcd/cat-exhaustiveness-spec.k                                                             | 61.44                                              | 84.7                  | 0.7253837072018889                                                    
| erc20/ds/transferFrom-failure-1-b-spec.k                                                  | 88.41                                              | 119.81                | 0.7379183707536933                                                    
| mcd/flipper-tau-pass-spec.k                                                               | 46.1                                               | 61.94                 | 0.7442686470778173                                                    
| bihu/forwardToHotWallet-failure-4-spec.k                                                  | 55.07                                              | 73.89                 | 0.7452970632020571                                                    
| benchmarks/requires01-a0gt0-spec.k                                                        | 39.25                                              | 52.46                 | 0.7481890964544414                                                    
| erc20/hkg/totalSupply-spec.k                                                              | 37.63                                              | 49.1                  | 0.7663951120162933                                                    
| mcd/vat-subui-pass-rough-spec.k                                                           | 75.2                                               | 97.24                 | 0.7733443027560676                                                    
| kontrol/test-expectreverttest-testfail_expectrevert_failandsuccess-0-spec.k               | 62.3                                               | 80.38                 | 0.7750684249813387                                                    
| erc20/ds/transferFrom-success-2-spec.k                                                    | 64.72                                              | 82.0                  | 0.7892682926829269                                                    
| mcd/vat-addui-pass-spec.k                                                                 | 78.04                                              | 98.54                 | 0.7919626547594886                                                    
| mcd/vat-slip-pass-rough-spec.k                                                            | 74.42                                              | 93.25                 | 0.7980697050938338                                                    
| bihu/forwardToHotWallet-failure-3-spec.k                                                  | 48.43                                              | 59.87                 | 0.8089193252046101                                                    
| mcd/vat-move-diff-rough-spec.k                                                            | 92.38                                              | 113.39                | 0.814710291912867                                                     
| examples/solidity-code-spec.md                                                            | 132.61                                             | 160.77                | 0.8248429433351994                                                    
| mcd/dstoken-approve-fail-rough-spec.k                                                     | 59.61                                              | 71.78                 | 0.8304541655057118                                                    
| erc20/ds/transfer-failure-1-b-spec.k                                                      | 61.3                                               | 73.77                 | 0.8309610952961909                                                    
| mcd/vat-deny-diff-fail-rough-spec.k                                                       | 62.6                                               | 75.24                 | 0.8320042530568847                                                    
| kontrol/test-arithmetictest-test_wmul_weakly_increasing_positive-uint256-uint256-0-spec.k | 62.27                                              | 74.84                 | 0.8320416889363976                                                    
| kontrol/test-allowchangestest-testallow_fail-0-spec.k                                     | 91.71                                              | 110.18                | 0.8323652205481937                                                    
| bihu/forwardToHotWallet-failure-2-spec.k                                                  | 43.2                                               | 51.33                 | 0.8416130917592052                                                    
| erc20/ds/transferFrom-failure-2-b-spec.k                                                  | 60.68                                              | 72.04                 | 0.8423098278734036                                                    
| mcd/vat-subui-pass-spec.k                                                                 | 134.07                                             | 157.65                | 0.8504281636536631                                                    
| erc20/ds/transferFrom-failure-1-c-spec.k                                                  | 69.7                                               | 81.16                 | 0.8587974371611632                                                    
| erc20/hkg/transferFrom-failure-1-spec.k                                                   | 54.81                                              | 63.78                 | 0.8593603010348072                                                    
| kontrol/test-allowchangestest-testallow-0-spec.k                                          | 37.56                                              | 43.62                 | 0.861072902338377                                                     
| erc20/ds/approve-success-spec.k                                                           | 45.01                                              | 52.19                 | 0.8624257520597816                                                    
| mcd/vat-subui-fail-rough-spec.k                                                           | 52.23                                              | 60.34                 | 0.8655949618826648                                                    
| kontrol/test-arithmetictest-test_wdiv_rounding-uint256-uint256-0-spec.k                   | 50.81                                              | 58.69                 | 0.8657352189470098                                                    
| erc20/ds/transfer-failure-2-a-spec.k                                                      | 48.24                                              | 55.16                 | 0.8745467730239305                                                    
| erc20/ds/transferFrom-failure-2-a-spec.k                                                  | 53.25                                              | 60.74                 | 0.8766875205795193                                                    
| mcd/vat-mului-pass-spec.k                                                                 | 69.94                                              | 79.76                 | 0.8768806419257772                                                    
| mcd/vat-muluu-pass-spec.k                                                                 | 142.47                                             | 160.56                | 0.8873318385650224                                                    
| kontrol/test-storetest-testaccesses-0-spec.k                                              | 39.91                                              | 44.88                 | 0.8892602495543671                                                    
| mcd/flopper-file-pass-rough-spec.k                                                        | 317.55                                             | 349.0                 | 0.9098853868194843                                                    
| kontrol/test-expectreverttest-testfail_expectrevert_false-0-spec.k                        | 59.42                                              | 64.93                 | 0.9151393808717079                                                    
| kontrol/test-owneruponlytest-testincrementasowner-0-spec.k                                | 66.38                                              | 72.39                 | 0.916977483077773                                                     
| mcd/functional-spec.k                                                                     | 452.64                                             | 489.79                | 0.924151166826599                                                     
| erc20/ds/approve-failure-spec.k                                                           | 48.69                                              | 52.63                 | 0.9251377541326239                                                    
| erc20/hkg/balanceOf-spec.k                                                                | 36.74                                              | 39.67                 | 0.9261406604487018                                                    
| mcd/end-subuu-pass-spec.k                                                                 | 62.37                                              | 67.06                 | 0.9300626304801669                                                    
| erc20/ds/transfer-success-1-spec.k                                                        | 53.9                                               | 57.53                 | 0.9369024856596558                                                    
| kontrol/test-expectreverttest-test_expectrevert_bytes4-0-spec.k                           | 76.77                                              | 81.86                 | 0.9378206694356217                                                    
| kontrol/test-expectcalltest-testexpectregularcall-0-spec.k                                | 38.82                                              | 41.39                 | 0.9379077071756463                                                    
| kontrol/test-emitcontracttest-testexpectemitdonotcheckdata-0-spec.k                       | 39.03                                              | 41.56                 | 0.9391241578440809                                                    
| bihu/forwardToHotWallet-success-1-spec.k                                                  | 41.51                                              | 44.13                 | 0.9406299569453885                                                    
| mcd/flipper-addu48u48-fail-rough-spec.k                                                   | 45.64                                              | 48.34                 | 0.9441456350848159                                                    
| kontrol/test-emitcontracttest-testexpectemit-0-spec.k                                     | 39.25                                              | 41.56                 | 0.9444177093358999                                                    
| erc20/ds/transfer-success-2-spec.k                                                        | 51.04                                              | 53.97                 | 0.9457105799518251                                                    
| benchmarks/structarg00-spec.k                                                             | 39.06                                              | 41.14                 | 0.9494409333981527                                                    
| kontrol/test-safetest-testwithdrawfuzz-uint96-0-spec.k                                    | 39.46                                              | 41.49                 | 0.9510725476018318                                                    
| bihu/forwardToHotWallet-failure-1-spec.k                                                  | 51.81                                              | 54.36                 | 0.9530905077262694                                                    
| benchmarks/staticloop00-a0lt10-spec.k                                                     | 38.41                                              | 40.26                 | 0.9540486835568802                                                    
| mcd/flipper-ttl-pass-spec.k                                                               | 46.28                                              | 48.5                  | 0.9542268041237114                                                    
| benchmarks/dynamicarray00-spec.k                                                          | 37.17                                              | 38.81                 | 0.9577428497809842                                                    
| benchmarks/overflow00-overflow-spec.k                                                     | 35.91                                              | 37.43                 | 0.9593908629441623                                                    
| kontrol/test-emitcontracttest-testexpectemitcheckemitter-0-spec.k                         | 38.98                                              | 40.62                 | 0.9596258000984736                                                    
| erc20/hkg/transfer-success-1-spec.k                                                       | 41.6                                               | 43.23                 | 0.9622947027527181                                                    
| kontrol/test-expectcalltest-testexpectstaticcall-0-spec.k                                 | 39.36                                              | 40.89                 | 0.962582538517975                                                     
| benchmarks/ecrecover00-sigvalid-spec.k                                                    | 43.37                                              | 45.04                 | 0.9629218472468916                                                    
| erc20/hkg/transfer-success-2-spec.k                                                       | 40.69                                              | 42.22                 | 0.9637612505921364                                                    
| erc20/ds/allowance-spec.k                                                                 | 41.96                                              | 43.5                  | 0.9645977011494253                                                    
| benchmarks/staticarray00-spec.k                                                           | 37.08                                              | 38.39                 | 0.9658765303464443                                                    
| mcd/vat-addui-fail-rough-spec.k                                                           | 65.37                                              | 61.89                 | 1.056228793019874                                                     
| erc20/ds/transferFrom-failure-1-a-spec.k                                                  | 75.91                                              | 69.79                 | 1.0876916463676742                                                    
| erc20/hkg/transferFrom-success-2-spec.k                                                   | 128.67                                             | 114.83                | 1.1205259949490551                                                    
| erc20/ds/transferFrom-success-1-spec.k                                                    | 90.19                                              | 79.84                 | 1.1296342685370742                                                    
| bihu/functional-spec.k                                                                    | 109.01                                             | 95.58                 | 1.1405105670642395                                                    
| mcd/dsvalue-read-pass-summarize-spec.k                                                    | 57.82                                              | 47.58                 | 1.2152164775115595                                                    
| kontrol/test-expectreverttest-test_expectrevert_returnvalue-0-spec.k                      | 60.07                                              | 48.31                 | 1.2434278617263506                                                    
| benchmarks/requires01-a0le0-spec.k                                                        | 52.22                                              | 37.92                 | 1.37710970464135                                                      
| erc20/hkg/allowance-spec.k                                                                | 52.53                                              | 37.89                 | 1.386381631037213                                                     
| erc20/hkg/transfer-failure-1-spec.k                                                       | 67.74                                              | 47.1                  | 1.438216560509554                                                     
| kontrol/test-arithmetictest-test_max1_broken-uint256-uint256-0-spec.k                     | 101.58                                             | 38.27                 | 2.6542984060621895                                                    
| TOTAL                                                                                     | 5251.62                                  | 5807.55    | 0.904274608053309